Clean electron main
